requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
autopep8==1.4.4
blessings==1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
curtsies==0.3.0
Django==2.2.3
django-appconf==1.0.3
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-decouple==2.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-js-asset==1.2.2
greenlet==0.4.15
idna==2.8
pilkit==2.0
Pillow==6.1.0
promise==2.2.1
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
stripe==2.32.1
typing==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
Whoosh==2.7.4
